requirements.txt
playwright==1.42.0
google-api-python-client==2.119.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
cryptography==42.0.5
python-dotenv==1.0.1